% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationData.R
\name{Annotation.filter}
\alias{Annotation.filter}
\alias{AnnotationData.filter}
\title{filter out/only include labels/categories presented in the input list}
\usage{
AnnotationData.filter(annotationData, ontologyData, labels = NULL,
  categories = NULL, include.labels = FALSE, include.categories = FALSE)
}

