% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationData.io.R
\name{AnnotationData.io.write}
\alias{AnnotationData.io.write}
\title{Write annotation data into mhealth folder structure and with mhealth filename convention.}
\usage{
AnnotationData.io.write(folder, annotationData, ontologyId = NA,
  annotatorId = NA, tz, gzip = TRUE, flatDir = FALSE, splitHour = TRUE,
  custom_name, append = FALSE, header = TRUE)
}
\arguments{
\item{folder}{the output folder}

\item{annotationData}{the input dataframe that matches mhealth specification.}

\item{ontologyId}{the ontology ID for this set of annotations}

\item{annotatorId}{the annotator's ID for this set of annotations}

\item{tz}{the time zone string (P/MHHMM) used in filename.}

\item{gzip}{whether to gzip the output csv file.}

\item{flatDir}{whether to use mhealth folder structure or just use flat directory.}

\item{splitHour}{whether to split input dataframe into hourly csv files.}

\item{custom_name}{if provided, the file name will be custom_name and all other file name preset parameters such as sensorType will be discarded.}

\item{append}{whether to append to a file if the file exists}

\item{header}{whether to add column header or not}
}
