% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationData.R
\name{AnnotationData.merge}
\alias{AnnotationData.merge}
\title{merge two or more annotation data frames and sort according to start time}
\usage{
AnnotationData.merge(annotationDataList, ...)
}
\arguments{
\item{annotationDataList}{list of annotation data frames}

\item{...}{other optional annotation data frames}
}

