% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationData.R
\name{AnnotationData.offset}
\alias{AnnotationData.offset}
\title{offset annotation data's start and stop timestamp by an offset value in seconds}
\usage{
AnnotationData.offset(annotationData, offsetValue = 0)
}
\arguments{
\item{annotationData}{annotation dataframe that matches mhealth specification.}

\item{offsetValue}{value in seconds specifies the offset time, could be negative, meaning go back to some time earlier. The default is 0, meaning no offset.}
}
\value{
annotation dataframe after timestamps being offset.
}
