% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationData.R
\name{AnnotationData.simplify.filter}
\alias{AnnotationData.simplify.filter}
\title{filter out/include only simplified annotation data frame with the presented labels}
\usage{
AnnotationData.simplify.filter(annotationData, labels, include.labels = TRUE,
  and.logic = TRUE)
}
