% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationData.R
\name{AnnotationData.simplify}
\alias{AnnotationData.simplify}
\title{simplify all annotations so that concurrent activities will be concatnated and timestamps will be connected}
\usage{
AnnotationData.simplify(annotationData)
}
