% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.extrapolate.v2.R
\name{extrapolate}
\alias{extrapolate}
\title{Apply extrapolate algorithm to a single stream of data. This is used as the inner function for `SensorData.extrapolate`.}
\usage{
extrapolate(t, value, range, noise_std, k = 0.1, spar = 0.3,
  confident = 0.5, fit_model = "spline")
}
\arguments{
\item{t}{input index sequence, start from 1}

\item{k}{duration of neighborhood to be used in local spline regression for each side, in seconds}

\item{spar}{between 0 and 1, to control how smooth we want to fit local spline regression, 0 is linear and 1 matches all local points. A good choice is 0.3 to penalize the maxed out points.}

\item{fit_model}{"spline", "linear"}
}
\note{
If wanting to obtain mediate results for plotting and debugging, please use this function.
}
\author{
Qu Tang
}
