% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyResponse.R
\name{FrequencyResponse.changeResolution}
\alias{FrequencyResponse.changeResolution}
\title{Change the frequency resolution of the input frequency response data}
\usage{
FrequencyResponse.changeResolution(frData, resolution = 0.01)
}
\arguments{
\item{frData:}{should be compatible with frequency response data format, with the first column be the frequencies, following by numeric columns
resolution: new resolution for frequencies. Default is 0.01Hz.}
}
\description{
If resolution is higher than the actual data resolution, it will do nothing,
if resolution is lower than the actual data, it will skip some of the data points evenly.
}

