% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyResponse.R
\name{FrequencyResponse.fft}
\alias{FrequencyResponse.fft}
\alias{FrequencyResponsed.fft}
\title{Compute frequency response for input sensor data.}
\usage{
FrequencyResponsed.fft(sensorData, Fs, type = "magnitude")
}
\arguments{
\item{sensorData}{input dataframe that matches mhealth specification.}

\item{Fs}{sampling rate of the input signal.}

\item{type}{"magnitude" or "vector". "magnitude": magnitude value of fft results; "vector": vector of fft results.}
}
\value{
frequency dataframe with HEADER `HEADER_FREQUENCY_STAMP,[FFTMAGNITUDE/FFTVECTOR]_[ORIGINAL_HEADER]`.
}

