% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Magnitude.R
\name{Magnitude.compute}
\alias{Magnitude.compute}
\title{Compute the magnitude value of sensor data}
\usage{
Magnitude.compute(sensorData)
}
\description{
Compute the magnitude value of sensor data
}

