% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MhealthFormat.R
\name{MhealthFormat.completionCheck}
\alias{MhealthFormat.completionCheck}
\title{Check the completion and report any found problems about a mhealth dataset}
\usage{
MhealthFormat.completionCheck(directory_path, subject_name_pattern,
  data_types_constraints)
}

