% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MhealthInteractiveVisualizer.R
\name{MhealthInteractiveVisualizer.getApp}
\alias{MhealthInteractiveVisualizer.getApp}
\title{The shiny app object of interactive visualizer for mhealth format data}
\usage{
MhealthInteractiveVisualizer.getApp(volumes = NULL)
}
\description{
This can be used if you want to embed the interactive visualizer in something else
}

