% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SamplingRate.R
\name{SamplingRate.plot}
\alias{SamplingRate.plot}
\title{Plot sampling rate over certain breaks (e.g. min, hour)}
\usage{
SamplingRate.plot(sr_dat, unit = "Count", ref)
}
\arguments{
\item{type:}{"Count" or "Hz" or "Dataloss".
"Count" will display the sample counts;
"Hz" will display the sampling rate in Hz (count divided by break time)
"Dataloss" will display the sampling rate in data loss percentage (count divided by reference)}
}
\description{
Plot sampling rate over certain breaks (e.g. min, hour)
}

