% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SamplingRate.R
\name{SamplingRate.summary}
\alias{SamplingRate.summary}
\title{Calculate sampling rate over a certain break (e.g. hour, min)}
\usage{
SamplingRate.summary(sensorData, breaks = "min")
}
\description{
Calculate sampling rate over a certain break (e.g. hour, min)
}

