% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.bokehplot}
\alias{SensorData.bokehplot}
\title{Plot sensor raw data using rbokeh interatively}
\usage{
SensorData.bokehplot(sensorData)
}
\arguments{
\item{sensorData}{input dataframe that matches mhealth specification.}
}
\description{
All columns will be plotted on the same graph with different colors.
}
