% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.createActigraphCsvHeader}
\alias{SensorData.createActigraphCsvHeader}
\title{create a character vector representing each line of the actigraph csv header}
\usage{
SensorData.createActigraphCsvHeader(startTime, downloadTime, samplingRate,
  sensorId, firmVersion, softVersion)
}

