% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.crop}
\alias{SensorData.crop}
\title{crop sensor data's y axis to a specific range}
\usage{
SensorData.crop(sensorData, range = NULL)
}
\arguments{
\item{sensorData}{input dataframe that matches mhealth specification.}

\item{range}{two values in vector specify upper and lower bound}
}
\value{
dataframe after being cropped
}
