% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.extrapolate.R
\name{SensorData.extrapolate}
\alias{SensorData.extrapolate}
\title{Apply extrapolate algorithm to sensor data. Algorithm was developed for activity count by NU mhealth group}
\usage{
SensorData.extrapolate(sensorData, lambda, interpolate, range, noiseStd)
}
\author{
Qu Tang
}

