% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.extrapolate.v2.R
\name{SensorData.extrapolate.v2}
\alias{SensorData.extrapolate.v2}
\title{Apply extrapolate algorithm to sensor data. Algorithm was developed for activity count by NU mhealth group.}
\usage{
SensorData.extrapolate.v2(sensorData, ...)
}
\arguments{
\item{sensorData}{input dataframe that matches mhealth specification.}

\item{range}{dynamic range, a vector with two values.}

\item{noise_std}{estimated noise standard deviation}

\item{sr}{sampling rate}

\item{k}{duration of neighborhood to be used in local spline regression for each side, in seconds}

\item{spar}{between 0 and 1, to control how smooth we want to fit local spline regression, 0 is linear and 1 matches all local points. A good choice is 0.3 to penalize the maxed out points.}

\item{fit_model}{"spline", "linear"}
}
\author{
Qu Tang
}
