% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.extrapolate.visualize.R
\name{SensorData.extrapolate.visualize}
\alias{SensorData.extrapolate.visualize}
\title{visualizing function for viewing the mediate results of the extrapolation algorithm}
\usage{
SensorData.extrapolate.visualize(extrapolated, input, groundtruth,
  dynamicRange = NULL, maxedOutRegion = NULL, regressionResults = NULL,
  showInput = TRUE, showGroundTruth = TRUE, showOutput = TRUE,
  showMarker = FALSE, showThreshold = FALSE, showEdges = FALSE,
  showCenter = FALSE, showNeighbors = NULL, showLines = NULL,
  showPoints = NULL)
}
\description{
---
title: SensorData.extrapolate.visualize
layout: launch
---
}
