% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.filter.R
\name{SensorData.filter.averageRemovalFIR}
\alias{SensorData.filter.averageRemovalFIR}
\title{Apply average removal FIR filter to the input sensor data frame each column over certain breaks (e.g. hour, sec, min and etc.)}
\usage{
SensorData.filter.averageRemovalFIR(sensorData, breaks, Fs, order)
}
\arguments{
\item{sensorData}{the input dataframe that matches mhealth specification.}

\item{breaks}{"sec","min","hour","day","week","month","quarter" or "year"; or preceded by integer and space.}

\item{Fs}{sampling rate of the input signal}

\item{order}{window size (in seconds) of filter}
}
\note{
If "breaks" is missing, filter will be applied on the whole sequence and return a list with a single dataframe.
}

