% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorData.filter.R
\name{SensorData.filter.butterworth}
\alias{SensorData.filter.butterworth}
\title{Apply high pass butterworth filter to the input sensor data frame}
\usage{
SensorData.filter.butterworth(sensorData, breaks, Fs, Fc, order)
}
\description{
Apply high pass butterworth filter to the input sensor data frame
}

