% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.filter.R
\name{SensorData.filter.resample}
\alias{SensorData.filter.resample}
\title{Apply bandlimited interpolation filter to the input sensor data frame each column over a certain break (e.g. hour, sec, min and etc.).}
\usage{
SensorData.filter.resample(sensorData, breaks, origSr, newSr)
}
\arguments{
\item{sensorData}{the input dataframe that matches mhealth specification.}

\item{breaks}{"sec","min","hour","day","week","month","quarter" or "year"; or preceded by integer and space.}

\item{origSr}{original sampling rate of each column}

\item{newSr}{the desired sampling rate for each column}
}
\value{
list of filtered dataframes.
}
\note{
If "breaks" is missing, filter will be applied on the whole sequence and return a list with a single dataframe.
}

