% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.generator.R
\name{SensorData.generator.fourierSeries}
\alias{SensorData.generator.fourierSeries}
\title{Generate simulated fourier series signal in mhealth format to simulate arbitrary periodical signal}
\usage{
SensorData.generator.fourierSeries(startTime, endTime, Fs, range, fbase, fstep,
  ampMin = 0, ampMax = 1, amp0 = 0, phaseMin = 0, phaseMax = 1,
  order = 3, noiseStd, parallel = FALSE, seed = 1)
}
\description{
If the signal arguments are vector, multiple columns will be generated
}

