% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorData.generator.R
\name{SensorData.generator.sinusoidal}
\alias{SensorData.generator.sinusoidal}
\title{Generate simulated sinusoidal signal in mhealth format}
\usage{
SensorData.generator.sinusoidal(startTime, endTime, Fs, range, f, amp, phase,
  noiseStd, seed)
}
\description{
If the signal arguments are vector, multiple columns will be generated
}

