% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.importActigraphCountCsv}
\alias{SensorData.importActigraphCountCsv}
\title{Import and convert Actigraph count csv files and load into data frame as in mhealth format.}
\usage{
SensorData.importActigraphCountCsv(filename)
}
\arguments{
\item{filename}{full file path of input Actigraph raw csv file.}
}
\seealso{
[`SensorData.importCsv`](SensorData.importCsv.html), [`SensorData.importGT3X`](SensorData.importGT3X.html), [`SensorData.importBinary`](SensorData.importBinary.html)
}

