% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.importBinary}
\alias{SensorData.importBinary}
\title{Import and decode binary file from the smart watch and load into dataframe as mhealth format.}
\usage{
SensorData.importBinary(filename, dest = file.path(getwd(), ".fromBinary"))
}
\arguments{
\item{filename}{full file path of input smart watch binary data file.}

\item{dest}{full directory path of destination folder. Default is ".fromBinary" folder of current working directory.}
}
\note{
It will call `SensorData.importCsv` after decoding.
}
\seealso{
[`SensorData.importCsv`](SensorData.importCsv.html), [`SensorData.importGT3X`](SensorData.importGT3X.html), [`SensorData.importActigraphCsv`](SensorData.importActigraphCsv.html)
}

