% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.importCsv}
\alias{SensorData.importCsv}
\title{Import mhealth sensor data file and load into memory as data frame in mhealth format}
\usage{
SensorData.importCsv(filename)
}
\description{
Import mhealth sensor data file and load into memory as data frame in mhealth format
}

