% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.interpolate}
\alias{SensorData.interpolate}
\title{Interpolate the missing points and unify sampling interval for the input sensor data}
\usage{
SensorData.interpolate(sensorData, method = "spline_natural",
  polyDegree = 3)
}
\description{
Interpolate the missing points and unify sampling interval for the input sensor data
}

