% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.io.R
\name{SensorData.io.write}
\alias{SensorData.io.write}
\title{Write sensor data into mhealth folder structure}
\usage{
SensorData.io.write(folder, sensorData, sensorType, dataType, sensorId,
  versionCode = "NA", tz, gzip = TRUE, flatDir = FALSE,
  splitHour = TRUE)
}

