% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.io.R
\name{SensorData.io.writeAsActigraphRaw}
\alias{SensorData.io.writeAsActigraphRaw}
\title{write sensor data frame into actigraph raw}
\usage{
SensorData.io.writeAsActigraphRaw(folder, sensorData, headerStr, custom_name)
}

