% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.offset}
\alias{SensorData.offset}
\title{offset sensor data's timestamp by an offset value in seconds}
\usage{
SensorData.offset(sensorData, offsetValue = 0)
}
\arguments{
\item{sensorData}{input dataframe that matches mhealth specification.}

\item{offsetValue}{value in seconds specifies the offset time, could be negative, meaning go back to some time earlier. The default is 0, meaning no offset.}
}
\value{
dataframe after timestamps being offset
}

