% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.parseActigraphCsvHeader}
\alias{SensorData.parseActigraphCsvHeader}
\title{parse actigraph csv header to get related version and sampling rate information}
\usage{
SensorData.parseActigraphCsvHeader(filename, header = TRUE)
}
