% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.summary.counts.R
\name{SensorData.summary.counts.compute}
\alias{SensorData.summary.counts.compute}
\title{Compute activity counts using [citation] algorithm}
\usage{
SensorData.summary.counts.compute(sensorData, breaks, filterType = "butter",
  cutoffs = c(0.6, 10), integrationType = "trapz")
}
\arguments{
\item{sensorData}{input sensor data frame that fits the mhealth specification}

\item{breaks}{epoch to compute counts on}

\item{filterType}{"butter", "bessel"}

\item{cutoffs}{cut off frequencies to be used in filtering, default is 0.6Hz and 10Hz. If choosing bessel, the low pass cut off would be multipled by 2 when being used.}

\item{integrationType}{the integration method to be used: "trapz", "power", "sum", "meanBySecond", "meanBySize"}
}

