% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaker.count.R
\name{shaker.count}
\alias{shaker.count}
\title{function to calculate the count values over series of epoches with specific count input parameters for shaker data}
\usage{
shaker.count(raw_data = "shaker1", epoches = c("2 sec", "5 sec", "10 sec",
  "15 sec"), resamples = c(0), filterTypes = c("butter"),
  low_bandwidths = c(0.25), high_bandwidths = c(5), intTypes = c("trapz"))
}
\arguments{
\item{raw_data}{the dataset name, if it's a string, the function will try to load it using `data()` function otherwise will take it directly as a shaker dataframe}

\item{epoches}{epoches in character vector that has format "num timeunit", such as "5 sec".}

\item{filterTypes}{filters to be used when computing count}

\item{low_bandwidths}{low end bandwidths of the filter}

\item{high_bandwidths}{high end bandwidths of the filter}

\item{intTypes}{integration types of the count algorithm: "trapz" or "power"}
}
\note{
This function would be only feasible for shaker type of data. There are two dataset attached with the package "shaker1" and "shaker2". Please use `?shaker1` or `?shaker2` to check its format.
}

