% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaker.plot.R
\name{shaker.plot}
\alias{shaker.plot}
\title{function to plot the output of walkrun.count function for walkrun data}
\usage{
shaker.plot(shaker_count, layout = SR ~ GRANGE, series = "DEVICE",
  sample_group = NULL, scale = "z-score", subset_condition = EPOCH ==
  "5 sec")
}
\arguments{
\item{shaker_count}{output dataframe of walkrun.count function}

\item{layout}{formula to specify the layout of subplots using column names, left side is rows and right side is columns. Please check the output of walkrun.count function to see all possible columns}

\item{series}{string to specify the column name to be used as color series on the same subplot}

\item{sample_group}{string to specify the column name to be used as shape series for sample points on the same subplot}

\item{scale}{scaling method to be used for each count type: "z-score" or "regression", or "none". "regression" is not supported by now.}

\item{subset_condition}{logical condition to subset the input dataframe before plotting it, same as the input to subset function of base R}
}
\note{
This function would be only feasible for the output of walkrun.count function of walkrun type of data. There are two dataset attached with the package "walkrun1" and "walkrun2". Please use `?walkrun1` or `?walkrun2` to check its format.
}

