% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shaker1}
\alias{shaker1}
\title{Accelerometer data collected using Actigraph GT3XBT and GT9X on a ecliptical shaker ranging from 240 RPM to 480RPM}
\format{A data frame with 10 variables:
\describe{
  \item{HEADER_TIME_STAMP}{timestamp for data sample, POXIct}
  \item{X_ACCELERATION_METERS_PER_SECOND_SQUARED}{calibrated acceleration value on x axis in g force, double}
  \item{Y_ACCELERATION_METERS_PER_SECOND_SQUARED}{calibrated acceleration value on y axis in g force, double}
  \item{Z_ACCELERATION_METERS_PER_SECOND_SQUARED}{calibrated acceleration value on z axis in g force, double}
  \item{LOCATION}{"Shaker", char}
  \item{SR}{sampling rate, char}
  \item{RPM}{RPM of the shaker in Hz, double}
  \item{DEVICE}{device id: MOS is for GT3XBT, TAS is for GT9X, char}
  \item{GRANGE}{G range of the device, char}
}}
\source{
\url{http://qutang.github.io}
}
\usage{
shaker1
}
\description{
A dataset containing raw segments of accelerometer data collected using GT3XBT (3 devices) and GT9X (1 device) on an ecliptical shaker ranging from 60RPM to 480RPM at sampling rate from 30Hz to 100Hz. The g range for both devices are +/- 8G. z-axis is facing up on the shaker.
}
\keyword{datasets}

