% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData.absoluteMean}
\alias{SummaryData.absoluteMean}
\title{Calculate summary value (absolute mean value) over a certain break (e.g. hour, min).}
\usage{
SummaryData.absoluteMean(sensorData, breaks)
}
\description{
If certain break is not provided or missing, will use the entire sequence
}

