% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData.auc}
\alias{SummaryData.auc}
\title{Calculate summary value (area under curve) over a certain break (e.g. hour, min)}
\usage{
SummaryData.auc(sensorData, breaks = "min")
}
\description{
Calculate summary value (area under curve) over a certain break (e.g. hour, min)
}

