% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData.ggplot}
\alias{SummaryData.ggplot}
\title{Plot summary data using ggplot2}
\usage{
SummaryData.ggplot(summaryData)
}
\arguments{
\item{summaryData}{input dataframe that matches mhealth sensor data format.}
}
\description{
All columns will be on the same graph with different colors.
}

