% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData.simpleMean}
\alias{SummaryData.simpleMean}
\title{Calculate summary value (simple mean) over a certain break (e.g. hour, min)}
\usage{
SummaryData.simpleMean(sensorData, breaks)
}
\arguments{
\item{sensorData:}{should be compatible with the mhealth sensor data format, first column should be HEADER_TIME_STAMP, and the following arbitrary number of columns should be numeric}
}
\description{
Calculate summary value (simple mean) over a certain break (e.g. hour, min)
}

