% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkrun.count.R
\name{walkrun.count}
\alias{walkrun.count}
\title{function to calculate the count values over series of epoches with specific count input parameters for walkrun data}
\usage{
walkrun.count(raw_data = "walkrun1", epoches = c("2 sec", "5 sec", "10 sec",
  "15 sec"), resamples = c(0), filterTypes = c("butter"),
  low_bandwidths = c(0.25), high_bandwidths = c(5), intTypes = c("trapz"))
}
\arguments{
\item{raw_data}{the dataset name, if it's a string, the function will try to load it using `data()` function otherwise will take it directly as a walkrun dataframe}

\item{epoches}{epoches in character vector that has format "num timeunit", such as "5 sec".}

\item{filterTypes}{filters to be used when computing count}

\item{low_bandwidths}{low end bandwidths of the filter}

\item{high_bandwidths}{high end bandwidths of the filter}

\item{intTypes}{integration types of the count algorithm: "trapz" or "power"}
}
\note{
This function would be only feasible for walkrun type of data. There are two dataset attached with the package "walkrun1" and "walkrun2". Please use `?walkrun1` or `?walkrun2` to check its format.
}

