% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walkrun2}
\alias{walkrun2}
\title{Walking and running accelerometer data collected using Actigraph GT3X at dominant waist}
\format{A data frame with 10 variables:
\describe{
  \item{HEADER_TIME_STAMP}{timestamp for data sample, POXIct}
  \item{X_ACCELERATION_METERS_PER_SECOND_SQUARED}{calibrated acceleration value on x axis in g force, double}
  \item{Y_ACCELERATION_METERS_PER_SECOND_SQUARED}{calibrated acceleration value on y axis in g force, double}
  \item{Z_ACCELERATION_METERS_PER_SECOND_SQUARED}{calibrated acceleration value on z axis in g force, double}
  \item{LOCATION}{body locations in Camel style, factor}
  \item{SR}{sampling rate, factor}
  \item{MPH}{speed in mile per hour, double}
  \item{WEIGHT}{carried weight in pounds, factor}
  \item{SUBJECT}{subject id, factor}
  \item{SESSION}{trial session for a subject, factor}
}}
\source{
\url{http://qutang.github.io}
}
\usage{
walkrun2
}
\description{
A dataset containing raw segments of accelerometer data collected using GT3X at dominant waist for different speeds. GT3X has G range +/-3G.
}
\keyword{datasets}

