% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationData.R
\name{AnnotationData.addToGgplot}
\alias{AnnotationData.addToGgplot}
\title{add annotation bars to an existing ggplot (most likely a sensor data plot)}
\usage{
AnnotationData.addToGgplot(p, annotationData)
}

