% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnotationData.R
\name{AnnotationData.merge}
\alias{AnnotationData.merge}
\title{merge two or more annotation data frames and sort according to start time}
\usage{
AnnotationData.merge(annotationDataList, ...)
}
\description{
merge two or more annotation data frames and sort according to start time
}

