% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/FrequencyResponse.R
\name{FrequencyResponse.fft}
\alias{FrequencyResponse.fft}
\title{Compute frequency response for input sensor data}
\usage{
FrequencyResponse.fft(sensorData, Fs, type = "magnitude")
}
\description{
Compute frequency response for input sensor data
}

