% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MhealthInteractiveVisualizer.R
\name{MhealthInteractiveVisualizer.run}
\alias{MhealthInteractiveVisualizer.run}
\title{An interactive visualizer for mhealth format data based on shiny}
\usage{
MhealthInteractiveVisualizer.run()
}
\description{
An interactive visualizer for mhealth format data based on shiny
}

