% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.extrapolate.R
\name{Sensor.extrapolate.singleColumn}
\alias{Sensor.extrapolate.singleColumn}
\alias{SensorData.extrapolate.singleColumn}
\title{Apply extrapolate algorithm to a single stream of data. This is used as the inner function for SensorData.extrapolate.}
\usage{
SensorData.extrapolate.singleColumn(sensorData, param)
}
\description{
If wanting to obtain mediate results for plotting and debugging, please use this function
}
\author{
Qu Tang
}

