% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.cleanup}
\alias{SensorData.cleanup}
\title{Clean up sensor data by removing invalid timestamps, according to matched level (e.g. year, month, day, hour, min, sec)}
\usage{
SensorData.cleanup(sensorData, level = "year", gt = NULL)
}
\description{
Clean up sensor data by removing invalid timestamps, according to matched level (e.g. year, month, day, hour, min, sec)
}
\note{
Make sure that the data frame is including timestamps
}

