% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.clip}
\alias{SensorData.clip}
\title{Clip sensor data according to the start and end time}
\usage{
SensorData.clip(sensorData, startTime, endTime)
}
\description{
Clip sensor data according to the start and end time
}
\note{
Make sure that the data frame is including timestamps
}

