% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.extrapolate.R
\name{SensorData.extrapolate}
\alias{SensorData.extrapolate}
\title{Apply extrapolate algorithm to sensor data. Algorithm was developed for activity count by NU mhealth group.}
\usage{
SensorData.extrapolate(sensorData, lambda = 1, interpolate, range, noiseStd)
}
\arguments{
\item{sensorData}{input dataframe that matches mhealth specification.}

\item{lambda}{parameter to regularize extrapolation: 0.1-10; default is 1. (the smaller the more confidence on points close to edges).}

\item{interpolate}{"linear", "spline_fmm", "spline_natural", "aspline_original", "aspline_improved".}

\item{range}{dynamic range.}

\item{noiseStd}{standard deviation of the white noise.}
}
\author{
Qu Tang
}
