% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.filter.R
\name{SensorData.filter.bessel}
\alias{SensorData.filter.bessel}
\title{Apply low pass bessel filter to the input sensor data frame each column over certain breaks (e.g. hour, sec, min and etc.)}
\usage{
SensorData.filter.bessel(sensorData, breaks, Fs, Fc, order)
}
\arguments{
\item{sensorData}{the input dataframe that matches mhealth specification.}

\item{breaks}{"sec","min","hour","day","week","month","quarter" or "year"; or preceded by integer and space.}

\item{Fs}{sampling rate of the input signal}

\item{Fc}{cut off frequency of bessel filter}

\item{order}{formula order of bessel filter}
}
\note{
If "breaks" is missing, filter will be applied on the whole sequence and return a list with a single dataframe.
}
