% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.generator.R
\name{SensorData.generator.fourierSeries}
\alias{SensorData.generator.fourierSeries}
\title{Generate simulated fourier series signal in mhealth format to simulate arbitrary periodical signal.}
\usage{
SensorData.generator.fourierSeries(startTime, endTime, Fs, range, fbase, fstep,
  ampMin = 0, ampMax = 1, amp0 = 0, phaseMin = 0, phaseMax = 1,
  order = 3, noiseStd, parallel = FALSE, seed = 1)
}
\arguments{
\item{startTime}{POSIXct date object for start timestamp.}

\item{endTime}{POSIXct date object for end timestamp, often with format startTime + numeric duration.}

\item{Fs}{sampling rate of the simulated device.}

\item{range}{dynamic range of the simulated device.}

\item{fbase}{simulated base frequency of the generated signal.}

\item{fstep}{simulated frequency step parameter of the generated signal.}

\item{ampMin}{minimum value of generated amplitude.}

\item{ampMax}{maximum value of generated amplitude.}

\item{amp0}{the DC component of the generated signal.}

\item{phaseMin}{the minimum initial phase of the generated signal.}

\item{phaseMax}{the maximum initial phase of the generated signal.}

\item{order}{the order of frequency components of the generated signal.}

\item{noiseStd}{the standard deviation of the noise of the generated signal.}

\item{parallel}{whether to use parallel computing when generating signal.}

\item{seed}{the random seed to be used when adding noise.}
}
\description{
frequency component will be generated with formula: `n * fstep * fbase`. `n` is nth order. Amplitude and phase will have uniform distribution between min and max values. Values will be cut off if beyonding the dynamic range.
}
\note{
If arguments (range, fbase, fstep, ampMin, ampMax, phaseMin, phaseMax, noiseStd) are vectors, multiple columns will be generated accordingly. But the vectors should have the same length.
}
