% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.ggplot}
\alias{SensorData.ggplot}
\title{Plot sensor raw data using ggplot2}
\usage{
SensorData.ggplot(sensorData)
}
\arguments{
\item{sensorData:}{should be compatible with the mhealth sensor data format, first column should be HEADER_TIME_STAMP, and the following arbitrary number of columns should be numeric}
}
\description{
Plot sensor raw data using ggplot2
}

