% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.importActigraphCsv}
\alias{SensorData.importActigraphCsv}
\title{Import and convert actigraph raw csv files and load into data frame as in mhealth format}
\usage{
SensorData.importActigraphCsv(filename)
}
\note{
Please make sure the actigraph raw csv file has timestamp included
}

