% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.importBinary}
\alias{SensorData.importBinary}
\title{Import and decode binary file from the smart watch and load into dataframe as mhealth format}
\usage{
SensorData.importBinary(filename, dest = file.path(getwd(), ".fromBinary"))
}
\description{
The default destination directory for the decoded file is stored in .fromBinary folder of current working directory
}

