% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.R
\name{SensorData.importGT3X}
\alias{SensorData.importGT3X}
\title{Import and decode GT3X files and load into dataframe as mhealth format}
\usage{
SensorData.importGT3X(filename, dest = file.path(getwd(), ".fromGT3X"),
  split = FALSE)
}
\description{
The default destination folder will be .fromGT3X in current working directory
}

