% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.io.R
\name{SensorData.io.write}
\alias{SensorData.io.write}
\title{Write sensor data into mhealth folder structure and with mhealth filename convention.}
\usage{
SensorData.io.write(folder, sensorData, sensorType = NA, dataType = NA,
  sensorId = NA, versionCode = "NA", tz, gzip = TRUE, flatDir = FALSE,
  splitHour = TRUE, custom_name, append = FALSE, header = TRUE)
}
\arguments{
\item{folder}{the output folder}

\item{sensorData}{the input dataframe that matches mhealth specification.}

\item{sensorType}{the sensor type string used in filename.}

\item{dataType}{the data type string used in filename.}

\item{sensorId}{the sensor ID string used in filename.}

\item{versionCode}{the version code string used in filename; default is "NA".}

\item{tz}{the time zone string (P/MHHMM) used in filename.}

\item{gzip}{whether to gzip the output csv file.}

\item{flatDir}{whether to use mhealth folder structure or just use flat directory.}

\item{splitHour}{whether to split input dataframe into hourly csv files.}

\item{custom_name}{if provided, the file name will be custom_name and all other file name preset parameters such as sensorType will be discarded.}

\item{append}{whether to append to a file if the file exists}

\item{header}{whether to add column header or not}
}
