% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorData.summary.counts.R
\name{SensorData.summary.counts.compute}
\alias{SensorData.summary.counts.compute}
\title{Compute activity counts using [citation] algorithm}
\usage{
SensorData.summary.counts.compute(sensorData, breaks, filterType = "butter")
}
\arguments{
\item{sensorData}{input sensor data frame that fits the mhealth specification}

\item{breaks}{epoch to compute counts on}

\item{filterType}{"butter", "bessel"}
}

