% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData.ggplot}
\alias{SummaryData.ggplot}
\title{Plot summary data using ggplot2}
\usage{
SummaryData.ggplot(summaryData)
}
\arguments{
\item{summaryData:}{should be compatible with the mhealth sensor data format, first column should be HEADER_TIME_STAMP, and the following arbitrary number of columns should be numeric}
}
\description{
Plot summary data using ggplot2
}

