% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData.simpleMean}
\alias{SummaryData.simpleMean}
\title{Calculate summary value (mean value) for each column over a certain time break (e.g. hour, min)}
\usage{
SummaryData.simpleMean(sensorData, breaks)
}
\arguments{
\item{sensorData}{input dataframe that matches mhealth sensor data format.}

\item{breaks}{could be "sec", "min", "hour", "day", "week", "month", "quarter" or "year"; or preceded by an interger and a space.}
}
\note{
If certain break is not provided or missing, will use the entire sequence. The column name (except for the first column) of output dataframe would be: [SUMMARY\_METHOD]\_INPUT\_HEADER\_NAME

If certain break is not provided or missing, will use the entire sequence
}
