% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData.simpleMean}
\alias{SummaryData.simpleMean}
\title{Calculate summary value (simple mean) over a certain break (e.g. hour, min)}
\usage{
SummaryData.simpleMean(sensorData, breaks = "min")
}
\description{
Calculate summary value (simple mean) over a certain break (e.g. hour, min)
}

